api/pyproject.toml
[tool.poetry]
name = "leannlsearch"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
requests = "^2.31.0"
openai = "^0.28.1"
weaviate-client = "^3.25.2"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
